requirements.txt
PyYAML >= 3.10
acstore >= 20230101
artifacts >= 20220219
cffi >= 1.9.1
dfdatetime >= 20221112
dfimagetools >= 20240301
dfvfs >= 20240115
dfwinreg >= 20240229
dtfabric >= 20230518
libbde-python >= 20220121
libcaes-python >= 20240114
libcreg-python >= 20200725
libewf-python >= 20131210
libexe-python >= 20131229
libfcrypto-python >= 20240114
libfsapfs-python >= 20220709
libfsext-python >= 20220829
libfsfat-python >= 20220925
libfshfs-python >= 20220831
libfsntfs-python >= 20211229
libfsxfs-python >= 20220829
libfvde-python >= 20220121
libfwnt-python >= 20210717
libfwps-python >= 20240225
libfwsi-python >= 20240301
libhmac-python >= 20230205
libluksde-python >= 20220121
libmodi-python >= 20210405
libphdi-python >= 20220228
libqcow-python >= 20201213
libregf-python >= 20201002
libsigscan-python >= 20230109
libsmdev-python >= 20140529
libsmraw-python >= 20140612
libvhdi-python >= 20201014
libvmdk-python >= 20140421
libvsapm-python >= 20230506
libvsgpt-python >= 20211115
libvshadow-python >= 20160109
libvslvm-python >= 20160109
libwrc-python >= 20220104
pytsk3 >= 20210419
xattr >= 0.7.2 ; platform_system != "Windows"
